GM saving stats.
Also, implemented a CSVLogger for more lightweight logging
You now set the wanted number of frames per episode et al when instantiating the GameManager
Agents must have a reset method
Modified experiment to use new GM standard
Modified a few agents to have reset. NB Sarsa just passes in the reset.

+++ evaluate_experiment.py
@@ -29,7 +29,7 @@ def evaluate(experiment_log):
 
     episode_stats = [m.groups()
                      for m in map(episode_stats_re.match, log) if m != None]
-    episode_stats = [(int(reward), int(duration))
+    episode_stats = [(int(reward), float(duration))
                      for reward, duration in episode_stats]
     episode_stats = pd.DataFrame(episode_stats, columns=[
                                  "total_reward", "duration"])
